Fix race condition in CLONE emulation
This did not affect glibc's implementation of pthreads, which
does not pass parameters to the child using registers on AArch32.

The child thread was reading the values of saved registers from
the parent's stack and dbm_thread structure after releasing the
lock, therefore allowing the parent to overwrite these values.
Modified to copy the values to the child's stack before releasing
the lock.