AArch32: redesign the inline hash table lookup and dispatcher
The inline hash table lookup was modified to restore the values
of scratch registers in a preamble inserted at the start of each
fragment. This greatly simplifies the dispatch stage. Furthermore,
this allows to achieve better performance without the complexity
of combining POPs and translated indirect branches. This is
similar to the implementation used on AArch64.

Calls to the dispatcher were modified to only store the context on
the stack, which avoids the issue of signal handler execution
overwriting thread_data->scratch_regs.