Rewrite signal handling. Delays async signals to allow TPC-SPC translation
This allows execution to proceed correctly when the code cache is
flushed during the execution of a signal handler, when the
signal frame PC is inspected or modified by the application and
when a signal is delivered while a trace is being built.

The context passed to the signal handler is not preserved.
For synchronous signals, the signal frame PC is not translated and
sigreturn is not allowed. However, most applications simply siglongjmp
from this type of handlers, which is supported.