ELF loader: load both the interpreter and the executable
This emulates the Linux loader in loading both the image of the
interepreter and the executable in memory before handing execution to
the interpreter. Previously, we were emulating explicit
calls to the interpreter, e.g. an application launched as:

$ ./dbm /bin/true

would appear to the loader to have been launched as:

$ /lib/ld.so /bin/true

That generally worked correctly, except for OpenJDK java which failed
to load .so dependencies.

Known issues: the interpreter of dynamically linked shared objects
won't load correctly. ASLR is still not used.