Fix race condition in thread creation
The completion of the thread init routine was signalled by the child by
setting the tid in its own structure. However, it was possible for the
child to terminate and free its thread private structure before the
parent thread was scheduled again, causing an invalid access when the
parent resumed, in child_data->tid.