Traces: initialise the branch_cache_status variable
This variable is used when unlinkign and restoring code fragments to
deliver signals. Without initialising it when appending a
fragment to a trace, an invalid value previously written by an aborted
trace could be used. If a signal was delivered during the initial
execution of this fragment, then the signal handler would proceed to
incorrectly unlink it.

Receiving a signal during that window is very unlikely in most cases.
However, it is a common scenario on NVIDIA Denver, which slows down the
initial execution of newly generated code to translate it, and thus
multiplies the chances that we'll hit this scenario.