A64: allow the scanner to copy unknown instructions
This allows execution to continue if an application is probing for
support for newer instructions / an unsupported extension and the
processor will throw a SIGILL, causing the application to not use
these instructions. If the application is directly attempting to
use unsupported extensions, then this could cause unintended
behaviour, however the message we print should make it obvious
what's happening.