accept some command line options (--help, --version)
Add author(), usage() and version() functions to support --help
and --version as command line options to MAMBO.

* --help executes usage(), which prints the short version of
version(), the usage prototype, and the list of plugins and watched
functions (if PLUGINS_NEW is defined).

* --version executes the full version of version(), which includes the
name a version of the compiler used to build the binary, and author().
author() shown a note about the authors and the license.

The version defaults to GIT_VERSION, otherwise set to 'untracked'.
It can be overriden through CLFAGS: CFLAGS='-DDBM_VERSION=\"v0.0.0\"' make clean all