Add translation of load-reseved and store-conditional
In contrast to ARM, an LR/SC sequence (code between LR and SC) is very
limited on RISC-V platforms. A maximum number of 16 instructions and
only a part of the base "I" and "C" instruction set is permitted. Since
additional loads and stores are also excluded, instrumenting an
instruction inside the sequence will most likely turn it into an
"unconstrained LR/SC loop" resulting in the trailing SC to always fail
on our test device. The ISA only garanties for "constrained LR/SC loops"
to succeed eventually.

A LR/SC loop may spread over 2 or more basic blocks which makes the
translation a little complex. For now, one scrach register is used to
save the original value read by LR and translate the loop to a mix of
software and hardware atomic sequence. The scrach register is hardcoded
to x31 (t6) which could interfere with a function that makes use of x31
and contains this translation, but it seems to work for the most
programs (luckily).