Add handler for SVE in A64 dispatcher
Renamed 'pop_neon' and 'push_neon' to
'pop_simd' and 'push_simd' respectively,
with an IFDEF selecting the corresponding
codepath.

However, there are a few issues yet to be ironed
out.

Firstly, on a system with a Vector Length (VL)
of more than 128-bits it doesn't work. This
has been tested with QEMU and natively on an
AWS Graviton machine(c7g.4xlarge). The dispatcher
for some reason takes the wrong path on a
'CBZ' instruction at around basic block 152.
This leads to early program termination with
a negative return status.

Secondly, it only works with a VL of 128-bits
on a dynamically compiled (target) binary.
Statically compiled ones just SEGFAULT.

Thirdly, having the instructions for pushing
and popping the predicate registers makes it so
that nothing works at all. Regardless of VL
and how the (target) binary was compiled
(dynamically or statically).