Add initial support for AT_RSEQ_*
In newer operating systems AT_RSEQ_* parameters are automatically
added to binaries. This does not mean that the application
uses rseq, but it is a possibility. Currently we do not support
rseq as it is a complex issue, so for now we allow AT_RSEQ_* to
be passed, however a warning is printed, as there are no
guarantees rseq will not be used by the binary. In most of the
cases however resq is not used and majority of binaries should run
without any issues.