Fix incorrect instructions names in ARM32 scanners
For unknown reasons some of the NEON instructions has been updated
in the PIE arm and thumb definitions without reflecting those
changes in the aarch32 scanners. As a result MAMBO does not compile
for 32-bit architectures. This commit adjusts the names in the
scanner to match defintions in PIE.wq