Fix incorrect instructions names in ARM32 scanners (#124)
* Fix incorrect instructions names in ARM32 scanners

For unknown reasons some of the NEON instructions has been updated
in the PIE arm and thumb definitions without reflecting those
changes in the aarch32 scanners. As a result MAMBO does not compile
for 32-bit architectures. This commit adjusts the names in the
scanner to match defintions in PIE.wq

* Reduce size of RESERVED_BRK_SPACE for A32

For unknown reasons when RESERVED_BRK_SPACE is set to 128 * PAGE_SIZE
MABMO segfaults on scanning the first instructions. The problem
first transpires after mmap and remmap involving RESERVED_BRK_SPACE
is executed. This fix reduces the reserved space back to PAGE_SIZE
for A32 - the value used before RISC-V merge. One of the possible
explnations is smaller virtual memory on A32 compared to A64 and RISC-V.